Modified filtering test
